Infrastructure/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
pyyaml==6.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
redis>=5.0.0
hiredis>=2.2.0
psycopg2-binary>=2.9.0
asyncpg>=0.28.0

